refactor(control-api): extract guard/yaml_diff from policies/mod.rs
